+++ Source/Instrument2A03.cpp
@@ -183,8 +183,8 @@ void CInstrument2A03::SaveFile(CInstrumentFile *pFile)
 			pFile->WriteInt(i);
 			const char *pName = pSample->GetName();
 			std::size_t NameLen = strlen(pName);
-			pFile->WriteInt(NameLen);
-			pFile->Write(pName, NameLen);
+			pFile->WriteInt((unsigned int)NameLen);
+			pFile->Write(pName, static_cast<UINT>(NameLen));
 			pFile->WriteInt(pSample->GetSize());
 			pFile->Write(pSample->GetData(), pSample->GetSize());
 		}
